boney template
The base CSS template which is built on a 12-column layout, with breakpoints built in for desktop > tablets > large phones > small phones. The columns start at % widths according to screen-size but adjust to a more mobile-friendly structure ( 1 > 2 > 4 columns).
This is intended to be used as a very basic starting point for any responsive web project - i.e. you may need to adjust widths/borders etc. to suit your particular needs! It is nothing particularly groundbreaking, but I have had a lot of mileage out of it, so I thought I would share!